Display "FG is running..." in main window instead of separate window
* The separate window is annoying because when switching from a
workspace with FlightGear on top and FGo! below, to another workspace
and back to the initial workspace, the "FG is running..." little window
appears on top of FlightGear, which prevents from controlling it until
one explicitely switches to FlightGear's window.

* This commit displays the "FG is running..." message in red in FGo!'s
main window (in a Label widget) instead of the separate InfoWindow. When
FlightGear is not running, this message is replaced by a "Ready"
indication in green.

* This new behaviour allows to interact with FGo! while FlightGear is
running, except for the things that rely on stuff started in
App.startLoops(). Among others, it makes it possible to modify the
configuration before the fgfs process is completely terminated.

* As a bonus, the new Label widget indicates the exit status of the last
FlightGear run, or the signal number in case the fgfs process was
terminated by a signal.

* While FlightGear is running, the "Run FG" button is greyed out.

* If the fgfs process can't be started for some reason (e.g., missing
executable or insufficient permissions), then that reason is displayed
in the pop-up window.