Fix problems related to fgfs process monitoring and Tk+threads interaction
* Use the 'threading' module instead of the low-level '_thread'.

* Don't use unsafe Tk calls from other threads than the Tk GUI thread.
In order to do this, the non-main thread (running _monitorFgfsProcess())
uses queues to transmit data to the main thread and virtual events to
notify it when there is something new in one of the queues. This way, we
don't need to do any polling. Even subprocess.wait() doesn't use a loop
with short sleeps, because we don't use its timeout parameter.

* Fix a small bug where FGo! would use 100% of one CPU core between the
time (presumably very short) where fgfs' stdout and stderr are both
closed and the time where the fgfs process is actually terminated.

* Forward the stdout and stderr of fgfs to FGo!'s stdout as soon as it
is available. It would be possible to forward fgfs's stderr to FGo!'s
stderr, but in order to do this, we would need:
  - either to use two more threads (4 in total: GUI thread, stdout and
    stderr readers, and one more thread for joining the two previous
    ones and then wait()ing for fgfs to terminate);
  - or to use one more thread + polling (eeek!);
  - or to use a completely different implementation such as asyncio.

With asyncio, everything can be done in the GUI thread, which is
interesting in some way, but this requires to integrate the asyncio
event loop into the Tkinter event loop, which is non-trivial. This has
been done at <https://github.com/fluentpython/asyncio-tkinter>,
unfortunately its licensing is reportedly unclear at the time of this
writing. Otherwise, it would be possible in principle and relatively
easy to do with two threads, one for the GUI and one running the asyncio
event loop (and coroutines, tasks, etc.). I don't think this would
improve the performance compared to the implementation from this commit,
however the interface would certainly be more responsive than with the
everything-in-a-single-thread implementation.

Putting both stdout and stderr in the same pipe as done from
subprocess.Popen() presents a very small risk of producing interleaved
output in case of huge lines (not-too-large writes to pipes are atomic).
The risk is so small and the consequences so insignificant that adding
more threads (or reducing performance by using polling) in order to fix
this doesn't seem to be worth it. The asyncio-based solutions wouldn't
add more threads, though. If someone wants to redirect fgfs' stdout and
stderr to *different* files using FGo!'s output, that would be an
elegant way.

* Print on stdout when fgfs is terminated, plus its exit status or the
signal number in case it was killed by a signal.

* Move App.run_button enabling/disabling logic from App._runFG() to
App.runFG() to make it easier to understand and maintain.