Only display the error message for invalid config when "Run FG" is clicked
* Don't display the message on every change of the Text widget (which
can be quite annoying), but only when "Run FG" is clicked.

* src/gui/mainwindow.py (App.updateCommand): set
self.lastConfigParsingExc to the exception obtained from
CondConfigParser, if any, otherwise to None. This way, the next time
App._runFG() is called, it will have access to all relevant details for
displaying a proper error message to the user.