Override Ctrl-F binding of standard widgets in the main window
* Before this commit, Ctrl-F pressed in a Text or Entry widget of the
main window used to both run FlightGear and advance the cursor to the
next character of the widget. This is because of the standard Tkinter
binding for Ctrl-F in these widgets (a class binding).

* In order to fix this, we define a mixin class PassShortcutsToApp and
two classes MyText and MyEntry inheriting from it that bind Ctrl-F to
App.onControlF_KeyPress() where the shortcut is centrally handled. Other
windows than the main window (e.g., Preferences) should continue to use
standard Tkinter widgets, since Ctrl-F is only supposed to run
FlightGear when the main window is active.