Makefiles: central management for PROGNAME, LANGUAGES, PROG_ICON_SIZES
* shared.mk: new file defining the aforementioned variables; this file
can be included from other Makefiles to avoid inconsistent definitions
of these variables.

* It would be nice to be able to do the same for the Python one-liner
that retrieves the FFGo version, but it is more tricky to do than simple
variable definitions. This is because this one-liner must be invoked
from a particular directory in order to work, which seems difficult to
ensure for shared.mk which is included from different directories.