Add Setuptools packaging
* In order to be able to install FFGo in editable/development mode (as
done with 'pip install --editable'), rename the top-level 'src'
directory to 'ffgo' (name of the top-level package as seen by Python).
This requires slight modifications to many files, unfortunately.

* The 'ffgo' name being used by the new top-level package directory,
also rename the 'ffgo' script in the top-level directory. The new script
is named 'ffgo-launcher.py'. It is only used to run FFGo from the source
directory, without installing it. Proper installations done with
'pip install FFGo' do have an 'ffgo' executable whose name is clearly
distinct from 'ffgo-launcher.py'. This should make it clear in bug
reports how FFGo was run, among the two ways that are now available.

* README.rst, docs/INSTALL/INSTALL_en, INFORMATION_FOR_TRANSLATORS,
INFORMATION_FOR_PACKAGERS: adapt to the new situation and improve.

* data: move the whole directory tree under the 'ffgo' Python package
directory, i.e. data/foo/bar is now available as ffgo/data/foo/bar. This
way, runtime data files are easily accessible via the pkg_resources API,
regardless of the platform. This includes *.mo l10n files. For FHS
compliance, a trivial patch modifying LOCALE_DIR in ffgo/constants.py
should work, at the expense of portability (OK for Linux distro
packages).

* ffgo/main.py: new file; application initialization is moved from
src/__init__.py to this new file, and a main() entry point is provided
for Setuptools.

* ffgo/main.py (run): due to proper use of the pkg_resources API, it is
not necessary anymore for FFGo to chdir() at program startup.

* ffgo/main.py: rename 'root' to 'master' for consistency with other
modules and remove the optional argument trick in run() to capture a
reference to the master object, as it has no use in FFGo (the only place
where run() is called being ffgo/main.py:main()).

* ffgo/main.py: rename early_tk_init() to earlyTkInit() for consistency
with the rest of the module (and yes, I prefer mixedCase, as it is
readable yet shorter than xx_yy_zz...).

* ffgo/misc.py: new file for utility functions/classes potentially used
in many places of the application. For now, it only contains a thin
abstraction layer on top of the pkg_resources API. This allows for
shorter and more readable code where resources are used, and decouples
the FFGo code base from the pkg_resources API.

* ffgo/config.py (Config._read): significant changes in order to adapt
to the use of resources as well as to better handle errors (not silently
ignore them...). The function is a bit stricter as to what it accepts
("" is not treated as None anymore); none of the places where it is
called seems to be affected. IOW, passing an empty config file name is
now considered a bug (unless the OS is able to read such a file!).

* Makefile: remove the 'dist' target, which is now redundant with
'setup.py sdist'.