Distribute images derived from the CC0 helicopter image under CC0 too
* The original intent was to distribute the images under the same
license as the rest of FFGo for unity, i.e. WTFPL version 2. I think it
was, and is still legal to do under most juridictions. However, legal
matters being what they are, the safest course of action IMHO is to
publish the derived images under the same license as the original image,
i.e. CC0 1.0 Universal.

* It should still be possible for most users, *depending on how the CC0
license is interpreted in their respective juridictions*, to relicense
the CC0 works under WTFPLv2 or whatever license they want---no matter
how pointless that might be.