Improve restoring of geometry upon reset ("reload config")
* ffgo/config.py (Config.update): the default assignment for some
settings was missing.

* Restore the main window geometry in ffgo/gui/mainwindow.py:App.reset()
instead of main.py:run(). Since App.reset(first_run=True) is run by the
App constructor, this should not change application initialization
(apart from making it clearer). However, the main window geometry will
now be restored (as the state and geometry of other windows) when doing
"reload config" if it was actually saved in the config file.