New option to save the main window position
* ffgo/gui/configwindow.py (widgetMisc): new check button to "Remember
the main window position", i.e. store the whole geometry
'widthxheight+x+y' into WINDOW_GEOMETRY instead of only the main window
size ('widthxheight'). This option is turned on by default, which is
consistent with the way the Command Window and Output Window are
handled. This allows users to completely save their setup for the three
main windows used in FFGo in config files.

* ffgo/config.py: new SAVE_WINDOW_POSITION config file parameter,
corresponding to the IntVar Config.saveWindowPosition. The check button
mentioned above is directly linked to this variable.

* ffgo/gui/mainwindow.py (App.saveWindowsGeometry): depending on the value
of Config.saveWindowPosition, save the whole geometry (new feature) or
only the position of the main window (previous behavior).

* ffgo/data/config/presets: document the change for WINDOW_GEOMETRY
(which can now be of the form 'widthxheight+x+y' in addition to the
already-supported 'widthxheight').

* ffgo/data/config/presets: document the new parameter
SAVE_WINDOW_POSITION.

* ffgo/data/help/help_en: explain precisely what is saved for which
window when one chooses "Save & Quit".