README.rst: remove the direct screenshot inclusions
* README.rst: since platforms such as GitHub and PyPI don't render the
images correctly, replace this part with a sentence pointing to the FFGo
home page, where the screenshots will be displayed as expected. Maybe
the images would be rendered properly if included in the Git repo and in
release tarballs, however doing so sounds like a waste of space and
users' bandwidth. So, for now, only the home page will have direct
screenshots inclusion in the HTML page generated from README.rst.