Improve parsing of XML files (parking/groundnet and scenarios)
* Use xml.etree.ElementTree instead of
xml.etree.ElementTree.ElementTree.

* Don't ignore OSError or AttributeError exceptions while parsing these
files.

* Check parking/groundnet files ending in '.xml', not 'xml' only.

* Let ElementTree open the XML files for proper encoding detection
(instead of assuming the encoding of the file corresponds to the user's
locale, which is likely to fail for non-ASCII XML files).

* Properly handle the cases where elt.find("foobar") returns None
(i.e., no direct <foobar> child found in elt).

* Use elt.iterfind() instead of elt.findall() when iterating over
child elements of elt.

* ffgo/config.py (_get_carrier_data): don't rely on the relative order
of the <name> and <parking-pos> child elements of a scenario's <entry>
element for the result. Gracefully handle the cases where the
<entry> or <parking-pos> elements have no <name> child element.

* ffgo/gui/mainwindow.py (App.read_parking): don't fail if the root
element of a parking file has no <parkingList> or <parkinglist> child.
Don't fail if a <Parking> element has no 'name' attribute. Don't return
empty strings in the resulting list. Don't use .split('"') on the 'name'
and 'number' attributes of <Parking>. It doesn't seem to make sense
here; I suppose this dates from a time where FGo! used to parse XML files
"by hand", without ElementTree.

* All this has been tested on two large collections of parking files,
with the following code in App.testStuff():

  for airport in self.config.airport_icao:
      logger.debug("Looking at airport {}, found parkings:".format(
          airport))
      parkings = self.read_airport_data(airport, "park")
      if parkings:
          logger.debug(*parkings, sep='\n')

  from pprint import pformat
  logger.debug(pformat(self.config._readScenarios()))

The resulting outputs for each of the two collections, before and after
this commit, have been compared. The only difference found was a parking
with an empty 'name' attribute and no 'number' attribute which, before
this commit, would result in an empty parking name in the popup list.