Small fixes with respect to file operations
* Be explicit about the encoding: using the user's locale is not wise
because 1) it might be too weak for what we need to write and 2) if the
user changes his locale after a file has been written, the next read
from this file is quite likely to fail. In the cases concerned here, the
files probably only contain ASCII data, therefore it is unlikely people
encountered the problem; but better be on the safe side.

* Don't close() files after the corresponding 'with open(...)'
statement.