Change the default for "Airport data source" to "Scenery"
* This reflects FlightGear's current behavior when processing --parkpos
and should provide much better parking data for people using TerraSync.
The old default didn't make use of the TerraSync data, which is a pity.

* As indicated in ffgo/data/help/help_en, this requires to have
downloaded data for the desired airport in one of the FG_SCENERY
components before parking positions can be shown for this airport with
the new default setting.