Interactively propose to change APT_DATA_SOURCE to the new default
* ffgo/config.py (Config): new ALREADY_PROPOSED_CHANGES config file
parameter used to record the questions the user has already seen. It
corresponds to Config.alreadyProposedChanges and contains a comma
separated list of "question identifiers".

* ffgo/gui/mainwindow.py (App.proposeConfigChanges): new method. For
now, it only proposes to change the airport data source to Scenery if it
is not already the case and if the suggestion hasn't already been made
to the user.

* ffgo/gui/mainwindow.py (App.__init__): call App.proposeConfigChanges()
and depending on its return value, reread or don't reread the config
file in the App.reset() call that follows.

* ffgo/data/config/presets: document the new parameter.