Enabled/disabled state for 'Update list of installed airports'
* This menu entry is only relevant when 'Show installed airports only'
is selected, i.e. when Config.filteredAptList (FILTER_APT_LIST) is true.
Enable or disable it as appropriate whenever this variable is changed.

* ffgo/gui/mainwindow.py (App.__init__): add Tkinter tracing of
Config.filteredAptList, with
App._updUpdateInstalledAptListMenuEntryState() as the callback.
Beware: this function is *not* an App *method*. I added the
'*args, **kwargs' stuff because it seems difficult to predict how many
arguments will be passed by Tkinter to the callback. 'self' and
'INDEX_OF_INSTALLED_APT_LIST_MENU_ENTRY' in
App._updUpdateInstalledAptListMenuEntryState() are caught as closures.

* ffgo/gui/mainwindow.py (App.reset): call
App._updUpdateInstalledAptListMenuEntryState() to enable or disable the
'Update list of installed airports' menu entry depending on the state of
Config.filteredAptList.