Clean up handling of airport, parking, runway and carrier settings
* Config.airport, Config.park, Config.rwy and Config.carrier are all
tkinter.StringVar instances, therefore the natural value for the default
setting, or the absence of a setting in these variables, is the empty
string. For a start, this is guaranteed not to conflict with any
legitimate value for the corresponding names (i.e., what is passed to
fgfs' --airport, --parkpos, --runway and --carrier options). The old
defaults used to be the 'None' or 'Default' string, depending on the
particular setting. Using the empty string instead allows shorter and
more unified code for all these settings.

* ffgo/gui/mainwindow.py: use context-dependent translations for the
"None" and "Default" special values used in the parking and runway
labels when Config.park.get() or Config.rwy.get() is the empty string.
This is important because the word "none", for instance, can have
different translations depending on what it applies to (in French:
aucun/aucune/aucuns/aucunes).

* ffgo/gui/mainwindow.py (App.updateRunwayAndParkingLabels): new method.

* ffgo/gui/mainwindow.py (App.updateAirport): don't attempt to translate
parking or runway names: 1) we don't extract them with xgettext and 2)
that would be a nightmare for translators. Use
App.updateRunwayAndParkingLabels() which does the right thing.

* ffgo/gui/mainwindow.py (App.reset): don't forget to update the runway
and parking labels based on the possibly-new values in Config.rwy and
Config.park. This is done using App.updateRunwayAndParkingLabels().

* ffgo/gui/mainwindow.py (App.__init__): no need to initialize
App.translatedRwy and App.translatedPark to anything else than
StringVar: the appropriate value will be set by
App.updateRunwayAndParkingLabels() called from App.reset(), which is
already called from App.__init__().