Airport search triggered by Tkinter observers
* This commit makes the airport search code simpler, shorter and more
efficient (more reactive and using less CPU). Instead of periodically
checking whether the airport search text has changed, one attaches an
observer to the corresponding StringVar and Tkinter automatically calls
it whenever the variable is modified.

* ffgo/gui/mainwindow.py (App.__init__): new StringVar stored in the
'airportSearchText' attribute of the App instance. It is linked to the
App.airportSearch MyEntry object. App.searchAirports() is attached to
this variable as an observer.

* ffgo/gui/mainwindow.py (App): airportSearchStart(),
airportSearchStop(), airportSearchUpdate() and focusAirportList() are
not needed anymore, and thus removed. Same thing for the
'airportSearchIsRunning' and 'old_airport_search' attributes.