Aircraft search triggered by Tkinter observers
* This commit is quite analogous to what was done very recently for the
airport search.

* It makes the aircraft search code simpler, shorter and more efficient
(more reactive and using less CPU). Instead of periodically checking
whether the aircraft search text has changed, one attaches an observer
to the corresponding StringVar and Tkinter automatically calls it
whenever the variable is modified.

* ffgo/gui/mainwindow.py (App.__init__): new StringVar stored in the
'aircraftSearchText' attribute of the App instance. It is linked to the
App.aircraftSearch MyEntry object. App.searchAircrafts() is attached to
this variable as an observer.

* ffgo/gui/mainwindow.py (App): aircraftSearchStart(),
aircraftSearchStop(), aircraftSearchUpdate() and focusAircraftList() are
not needed anymore, and thus removed. Same thing for the
'aircraftSearchIsRunning' and 'old_aircraft_search' attributes.