New feature: copy FG shell-equivalent command to clipboard
* ffgo/gui/mainwindow.py (App.copyFGCommandToClipboard): new method. On
X11, the Tkinter clipboard functions used here seem to use the PRIMARY
selection. The UTF8_STRING target type seems to be vaguely supported by
Tk's "clipboard get" command, but I didn't manage to use it successfully
with Tkinter's clibpoard_append() using Python 3.4.3 and Tcl/Tk 8.6.
Therefore, this new method can only copy FG commands that can be encoded
to ISO 8859-1. An error dialog is displayed if this is not possible.

* ffgo/gui/mainwindow.py (App.__init__): register the new function under
the Tools menu.