Add airport parking tooltips in the parking popup
* The newly-added tooltips contain most of the metadata available for
parking positions (i.e., everything pilots are likely to be concerned
about, but not FlightGear implementation details such as the parking
'index' or its 'pushBackRoute'). The parking name/number its associated
flight type are not repeated in the tooltip since they are already
displayed in the parking popup.

* ffgo/gui/mainwindow.py (App.airportParkingTooltip,
App.populateAirportParkingPopup): new methods.
populateAirportParkingPopup() populates the popup menu with the usual
parking names as well as new tooltips, made with tooltip.ArrayToolTip.
populateAirportParkingPopup() delegates the preparation of each parking
tooltip text to airportParkingTooltip().

* ffgo/gui/mainwindow.py (App.popupPark): move and extend the code
related to airport parking positions to the new method
populateAirportParkingPopup().