Read parking positions from apt.dat.gz as a fallback
* If the standard lookup from the groundnet files gives no result, look
into apt.dat.gz. This is slow, especially when the ICAO is far in the
alphabet, but the last 50 results are cached in memory to mitigate the
inconvenience.

* Don't be impatient after clicking on the parking button, the popup
will appear. Future commits will add visual feedback.

* ffgo/fgdata/apt_dat.py: new parser for apt.dat/apt.dat.gz. Should be
compliant with the v1000 spec. Currently, it is only used to lookup
fallback parking positions for a given airport.