InfoWindow improvements
* ffgo/gui/infowindow.py (InfoWindow.__init__): new 'title' keyword
argument to allow choosing the window title. If passed as None (the
default), PROGNAME is used.

* ffgo/gui/infowindow.py (InfoWindow.__init__): remove the 'cnf'
argument; don't pass it to Toplevel.__init__(), as it sometimes causes
problems and doesn't conform to the Tkinter documentation I found.

* ffgo/gui/infowindow.py (InfoWindow.__init__): call update_idletasks()
followed by grab_set() at the end.

* ffgo/gui/infowindow.py (InfoWindow.getFirstLine): method now unused,
remove it.

* ffgo/config.py (_ProcessApt.make_window): don't call
self.window.update(). The update_idletasks() in InfoWindow.__init__()
seems to be sufficient and Tkinter's update() seems to be a "last
resort" method, not warmly recommended.

* ffgo/gui/mainwindow.py (App.filterAirports): for the same reason,
don't call infoWindow.update().