Add visual feedback during the fallback apt.dat.gz lookup for parkings
* Create a modal window (using the InfoWindow class) during the
apt.dat.gz lookup for parkings when nothing was found in the groundnet
file. This will let the user know he just has to be patient until the
parking popup menu appears.

* Move the time-consuming apt.dat.gz lookup to a separate worker thread
to keep the interface responsive. Use a queue for communication between
this thread and the GUI thread, plus a Tkinter virtual event to allow
the worker thread to wake up the GUI thread when a new item has been put
on the queue. Transmit exceptions caught in the worker thread (for
instance due to apt.dat.gz parsing errors) to the GUI thread via the
queue, and raise them from the event handler (which runs in the GUI
thread) so that they are handled just as other exceptions raised from
the GUI thread.