Remove FG_HOME and DEFAULT_TERRASYNC_DIR
* Apparently, the HOME environment variable is not always defined on
Windows. This prevents FFGo from starting in such a case
(<https://sourceforge.net/p/flightgear/mailman/message/34584497/>).

* Besides, the correct value of FG_HOME depends on the FlightGear
version, which makes things tricky in the Settings dialog (and maybe
confusing for users if the program tries to be too clever), since the
fgfs executable, and thus the FlightGear version we can detect, can be
changed in the same tab as where Config.FG_scenery is defined.

* DEFAULT_TERRASYNC_DIR depends on FG_HOME.

* In the current state, a non-empty default value for Config.FG_scenery
prevents setting an empty value for this parameter, which I find
undesirable (since an empty value allows to omit the --fg-scenery option
in the fgfs command line, which may be a legitimate goal to achieve).

* For all these reasons, this commit removes FG_HOME and
DEFAULT_TERRASYNC_DIR from ffgo/constants.py as well as the default
setting to DEFAULT_TERRASYNC_DIR for Config.FG_scenery. The default
value for Config.FG_scenery is now empty, as it used to be some time
ago...