New apt.dat parser; apt digest file format with index; airport tooltip
* Replace the apt.dat parser in ffgo/config.py with a new parser located
in ffgo/fgdata/apt_dat.py. This new parser should be compliant with the
APT1000 specification, provide better error reporting and fix a few
small bugs (such as slightly incorrect calculations of v810 airport
coordinates [two opposite runway ends counted as one] or computation of
runway names from v810 data ["6" instead of "06" for the "other
runway" opposed to "24"]).

It also extracts more information about airports and runways, unifies
recently added extraction of parking metadata from apt.dat with this
airport metadata, and generates an "apt digest file" (the file whose
path is given by constants.APT) in a new format bringing the following
improvements:

  - more robust (proper header including magic number, encoding
    declaration and format version; fields separated by the NUL char
    ('\0') instead of '=');

  - contains only the information FFGo must read on startup in order to
    make this startup as fast as possible; in particular, runway names
    for *all* airports are not read anymore on startup and held into
    memory, they are loaded on demand when needed for the selected
    airport;

  - since runway and parking metadata are loaded on demand from apt.dat
    (after trying the groundnet file in the case of parkings), the new
    format includes an index for each airport allowing to quickly find
    its data in apt.dat.gz. Of course, this requires keeping the apt
    digest file in sync with apt.dat.gz, which should not be much of a
    problem since the default for “Airport database update” is
    “Automatic” since version 1.3.0.

* The creation of the apt digest file is a bit slower than before (about
25 seconds on my system, as opposed to 10 or 11 before), but this should
only be necessary when apt.dat.gz changes---which does not happen very
often, the current version in FG 3.7 dating from 2013---or if I have to
again change the format of the apt digest file.

The slowdown is essentially caused by gzip.GzipFile.tell() being slow,
however this is necessary to create an efficient index. Using line
numbers as indices would be faster for creating the apt digest file (as
in, 11-12 seconds instead of 25), but then accessing the indexed data
would be much slower (as in, 2.3 seconds instead of 0.8). This would be
clearly undesirable, since such lookups are likely to be much more
frequent than rebuilding of the apt digest file---not to mention that
users would be confused with the interface taking about 2 seconds to
react to a click, would be likely to click again, etc.
[tests done with Python 3.4.3 on Linux 4.2.5 x86_64]

* When FFGo is started, or when it is asked to reload its config file,
the apt digest file is read and turned into a dictionary
(Config.airports) whose keys are ICAO codes and values corresponding
AirportStub instances. The AirportStub class is optimized for using just
the memory needed to store:

  - the ICAO;

  - the full airport name;

  - the latitude and longitude of its "centroid", which are needed
    globally to be able to find the nearest METAR station for a given
    airport, as well as to compute the set of "installed airports" (at
    least with the way Config._findInstalledApt() currently works);

  - the aforementioned index for fast access to detailed airport data in
    apt.dat.gz.

The Config.airport_icao, Config.airport_name and Config.airport_rwy
lists are not needed anymore and thus removed, since they would be
redundant with Config.airports and the data it gives access to (see
below).

* In order to read detailed airport data, a new method,
App.readAirportData(), is provided that returns Airport instances (not
AirportStub) and caches the last 50 Airport instances fetched in the
Config.aptDatCache deque for even faster access (no need to seek through
apt.dat.gz if the airport we are interested in is already in the cache).

* In order to be able to efficiently find the index of a given airport
in the airport list (which is needed to implement App.getIndex()),
App.updateAirportList() (formerly Config.updateAptLists()) and
App.buildAirportList() maintain the new App.shownAirports list always
synchronized with the contents of the App.airportList widget, similarly
to how App.shownAircrafts is kept in sync with the displayed aircraft
list. The elements of App.shownAirports are the AirportStub instances
from Config.airports' values corresponding to the airports currently
present (after applying the eventual search filter) in the airport list
widget (App.airportList).

* Startup and config reloading (precisely: Config._readApt()) should be
significantly faster when the airport list is filtered thanks to
Config._readInstalledAptSet() (formerly Config._readInstalledAptList())
now returning a frozenset.

* Try to consistently use the expression "apt digest file" to refer to
the file specified by constants.APT (i.e., %APPDATA%/FFGo/apt on
Windows, ~/.ffgo/apt on other OSes). Accordingly, rename
Config._makeApt() to Config._makeAptDigest() and _ProcessApt._makeApt()
to _ProcessApt._makeAptDigest().

* Remove Config.readCoord(), not needed anymore as the corresponding
data is readily available in Config.airports.

* Move Config.updateAptLists() to App.updateAirportList(): it makes more
sense there, since its main purpose is to generate the list of airports
that can be shown in the App.airportList widget, possibly filtered to
only include "installed airports". Accordingly, this method is not
called from Config.update() anymore but from App.resetLists() and
App.filterAirports().

* Add a colored, disabled (i.e., not clickable) and empty header to the
runway popup menu. This makes it more visible, visually similar to the
parking popup, and avoids it disappearing in a flash with the first
entry being accidentally selected if the user just clicked without
holding the button down.

* Add tooltips to the airport list giving basic information about the
airport under the mouse pointer (type of airport [land airport, seaplane
base or heliport], latitude, longitude, elevation and, if applicable,
the lists of land runways, water runways and helipads).

* ffgo/fgdata/apt_dat.py: many changes to the AptDat class (the new
apt.dat parser, etc.). New AptDatDigest class to help reading and
writing the apt digest file (note: most of the writing logic is in
AptDat.makeAptDigest(), since AptDat is the class giving access to the
data source for the apt digest file).

* ffgo/gui/mainwindow.py: remove the InfoWindow message and separate
thread for parking lookup in apt.dat (and corresponding
AptDatParkingLookupMsgType enum, related methods...). This complexity is
not justified anymore, because the lookup is now fast enough thanks to
the index stored for each airport in the apt digest file (about 0.8
seconds in the worst case on my computer, for an airport whose data is
located at the end of apt.dat.gz, and instantaneous after the first
access for any airport, thanks to Config.aptDatCache).