More airport and runway metadata
* ffgo/fgdata/airport.py: add a number of classes and enums to precisely
describe runway metadata (including data in the old v810 format).

* ffgo/fgdata/apt_dat.py (AptDat): changes to handle the new metadata.

* ffgo/gui/mainwindow.py: add tooltips in the runway popup menu.

* Print airport elevation in meters as well as in feet in the airport
tooltip.

* ffgo/fgdata/apt_dat.py (AptDat.testAllAirports): new method largely
based on AptDat.makeAptDigest() to allow automated, thorough reading of
all airports with their runway data at once from apt.dat (printing the
runway tooltips to a file). It reveals 35 bugs in apt.dat related to the
buoys flag for water runways, and all valid uses of this flag set it to
0 (no buoys). For this reason, treatment of this metadata is commented
out, and therefore the buoys flag can't be seen in water runway
tooltips.