Implement Vincenty's formula + fallbacks for the inverse geodesic problem
* ffgo/geo/geodesy.py: new file essentially containing an implementation
of Vincenty's formula as well as an abstraction layer
(GeodCalc.inverse()) to do the corresponding computations with the best
implementation available automatically chosen.

* ffgo/geo/geodesy.py (NVector): simple class implementing n-vectors.

* ffgo/geo/geodesy.py (EarthModel): class containing constants from the
WGS 84 model of the Earth, plus a method to compute the Gaussian radius
of curvature for a given latitude.

* ffgo/geo/geodesy.py (GeodCalc): class for performing basic geodesic
calculations.

* ffgo/geo/geodesy.py (GeodCalc.vincentyInverseWithFallback): method
using fallback strategies when Vincenty's algorithm doesn't work
(ellipsoidal Earth model projected to a plane for short distances,
spherical approximation for long equatorial lines and to determine
azimuths). The only case this algorithm can't handle (assuming the
various thresholds have been given appropriate values) is that of
antipodal or nearly antipodal endpoints. For this case, when it can be
triggerred in FFGo (which is not the case ATM), users can install
GeographicLib. This is suggested in an explanatory message when this
situation is encountered.

* ffgo/fgdata/apt_dat.py: adapt to use the new abstraction layer
(GeodCalc.inverse()).