Add GeodCalc.modifiedFccDistance()
* This new method is a standalone implementation of the idea used in the
previous commit: instead of fccDistance(), use spherical approximation
with the Gaussian radius of curvature if the path endpoints are not too
far away from one of the poles. In other cases, the distance is computed
with fccDistance().

* modifiedFccDistance() is not used in vincentyInverseWithFallback()
even though that would make the code shorter. This is because that would
cause a useless, duplicate computation of the angle between the two
n-vectors.