vincentyInverseWithFallback(): return azimuth 0.0 for identical points
* I had previously used -180.0 for this degenerate case, believing that
was what GeographicLib's Geodesic.WGS84.Inverse() was doing in such a
case; but actually, it doesn't always behave this way, so let's just use
0.0 to avoid readers think there is some special reason this -180.0
value was chosen for. Anyway, the value is *meaningless* in this case
(identical start and end points), and should not be used at all by
callers!