Correct calculation to find the nearest METAR station
* ffgo/gui/metar.py (Metar._nearestMetar): the previous code didn't
behave properly for points located on either side of the ±180° meridian,
or close to one of the poles. Use a fast method
(GeodCalc.modifiedFccDistance()) that gives reasonably correct results
to find the best 15 candidates for the title of "closest station", then
find the closest among these using GeodCalc.inverse().

* For information, here are the durations in seconds of the first phase
on my computer[1], with different methods to evaluate the distance
between the selected airport and each candidate station:

  Previous method (may give incorrect results in some cases) 0.022
  GeodCalc.modifiedFccDistance                               0.037
  GeodCalc.vincentyInverseWithFallback with 1e-6 precision   0.126
  GeodCalc.vincentyInverseWithFallback with 1e-12 precision  0.148
  GeodCalc.karneyInverse                                     1.045

[1] For 5099 METAR stations (list read from metar.dat.gz)