Be more tolerant with the format version line of apt.dat
* ffgo/fgdata/apt_dat.py (AptDat._formatLine_cre): apparently, apt.dat
files generated by WED (distributed at
<http://gateway.x-plane.com/NOTAMs>) have a format version line that the
previous regexp would not accept. FlightGear doesn't ship such files as
far as I know, therefore users of current versions of FlightGear are not
affected. This change is mainly intended to avoid problems in the
future, as well as to allow experiment with recent apt.dat files.