Add App.selectNewAirport()
* ffgo/gui/mainwindow.py (App.updateAirport): new keyword argument
'allowLeavingCarrierMode', defaulting to True. When it is true,
App.resetCarrier() is called if the airport selected in the airport list
is different from self.config.airport. This fixes what I consider a
little bug: previously, when in carrier mode, clicking on a different
airport in the airport list didn't really select it, as it didn't leave
carrier mode.

* ffgo/gui/mainwindow.py (App.selectNewAirport): new method to
programmatically select a new airport, leaving carrier mode if
necessary.