New file: ffgo/gui/widgets.py -> MyTreeview, SortOrder, Column, AirportChooser
* ffgo/gui/widgets.py (MyTreeview): new class derived from Ttk's
Treeview, offering better handling of keyboard events.

* ffgo/gui/widgets.py (SortOrder): new enum describing ascending or
descending sort order.

* ffgo/gui/widgets.py (Column): new class used to hold column metadata
for Treeview widgets.

* ffgo/gui/widgets.py (AirportChooser): new class used to implement a
hopefully convenient airport chooser based on a Ttk Treeview, an Entry
and a Button. This is similar to what already exists in FFGo's main
window, but with better presentation (a real multicolumn list/tree
widget) and more power (the table can be sorted by clicking on any
column header, in ascending or descending order). Last but not least,
this new "abstract widget" is smarter about selecting a suitable item in
the list whenever the search field is updated. The general idea is to:
  - select the airport whose code is equal to the contents of the search
    field (ignoring case), if such an airport exists. This way, typing
    "lfpo" in the search field automatically selects Paris Orly (LFPO),
    not any airport containing this string in its name (such as Gulfport
    Biloxi Intl or Gulfport Jail)---those are still listed and available
    for selection. This is particularly convenient for airports such as
    Turtle Island, whose code "TTL" is contained in the name of many
    other airports (most of which containing the word "little");
  - otherwise, select the airport that was already selected before the
    last change to the search field;
  - if that is not possible (because this airport is not in the refined
    list), select the first airport from the list.

Currently, the widget only contains two columns (ICAO and airport name),
but it is very easy to add more. The only thing needed for that is to
include the corresponding data into the apt digest file, so that reading
it for the 34074 airports currently defined in apt.dat is fast enough.

This widget also uses a timer to delay computing of the updated list
after a change to the search field, in order to improve responsivity
when typing or erasing "fast" (same idea as in FGo!, but implemented in
a CPU-efficient way thanks to Tkinter observers).